+++ src/plugins/intellij/api/index.ts
@@ -4,4 +4,5 @@ export * from './intellijHttpRequest';
 export * from './intellijHttpResponse';
 export * from './intellijRandom';
 export * from './intellijVariables';
+export * from './intellijWindow';
 export * from './stubs';