+++ src/plugins/intellij/api/stubs/http-declare.ts
@@ -1,5 +1,5 @@
 import { HttpClient, HttpClientRequest, HttpResponse } from './http-client';
-import { jsonPath } from './http-client.common';
+import { CommonWindow, jsonPath, xpath } from './http-client.common';
 import { CryptoSupport } from './http-client.crypto';
 import { PreRequestHttpClientRequest } from './http-client.pre-request';
 
@@ -8,6 +8,8 @@ export interface IntellijJavascriptGlobal extends Record<string, unknown> {
   client: HttpClient;
   crypto: CryptoSupport;
   jsonPath?: jsonPath; // not implemented
+  xpath?: xpath; // not implemented
+  Window: CommonWindow;
   $exampleServer?: string; // not implemented
   /**
    * The object holds information about HTTP Response.